added some extra functionality

+++ src/main/java/net/codejava/spring/services/MeetingService.java
@@ -6,10 +6,10 @@ import net.codejava.spring.model.Meeting;
 
 public interface MeetingService {
 	public Meeting create(Meeting shop);
-    public Meeting delete(int id);
+    public Meeting delete(long id);
     public List<Meeting> findAll();
     public Meeting update(Meeting shop);
-//    public Employee findById(int id);
+//    public Meeting findById(long id);
 //    public List<Employee> findByFirstname(String name);
 //    public List<Employee> OrderByPanNumber();
 //	List<Employee> OrderByIdDescending();
